requirements.txt
# YouTube and video processing
google-api-python-client>=2.100.0
yt-dlp>=2025.1.0
ffmpeg-python>=0.2.0

# Transcription - MLX Whisper for Apple Silicon
mlx-whisper>=0.4.0

# LLM
ollama>=0.1.6
groq>=0.4.0

# Telegram
python-telegram-bot>=20.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
apscheduler>=3.10.0
pydantic>=2.0.0
loguru>=0.7.0
tenacity>=8.2.0

# Data handling
aiosqlite>=0.19.0
